requirements.txt
# Core web framework
Flask>=2.3.0

# Authentication
Flask-Login>=0.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Environment variables
python-dotenv>=1.0.0

# AI/ML integration
openai>=1.0.0

# WSGI server for production
gunicorn>=21.0.0

# Database ORM and drivers
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Additional utilities (if needed in future)
# requests>=2.31.0
# numpy>=1.24.0requests>=2.31.0
